First major wave of SDP development.
Makefile.ocpi.common: stdc_LIMIT/FORMAT macros in CXXFLAGS
adapt/os/ocpios/interfaces/include/OcpiOsDebug.h: add logGetLevel
adapt/os/ocpios/linux/src/OcpiOsDebug.cxx: ditto
adapt/os/ocpios/interfaces/include/OcpiOsSocket.h: add "all" option to read all requested
adapt/os/ocpios/linux/src/OcpiOsSocket.cxx: ditto
adapt/os/ocpios/linux/src/OcpiOsEther.cxx: warnings
adapt/os/ocpios/linux/src/OcpiOsSocket.cxx: recv can optionally persist until all requested.
-- send can persist until all written
components/file_read.hdl/file_read.vhd: more complete signal driving, better messages
components/file_write.hdl/file_write.vhd: better messages
core/application/src/OcpiServer.cxx: warning
core/container/include/ContainerWorker.h: string option to getpropertybytes to null-terminate.
core/container/src/ContainerWorker.cxx: ditto
core/dataplane/rdma_driver_interface/interfaces/src/DtSharedMemoryInternal.cxx: warning
core/dataplane/rdma_driver_interface/interfaces/src/DtTransferInternal.cxx: warning
core/dataplane/rdma_drivers/datagram/src/DtDataGramEtherDriver.cxx: warning
core/dataplane/rdma_drivers/datagram/src/DtDataGramUDPDriver.cxx: warning
core/dataplane/rdma_drivers/dma/src/DtDmaXfer.cxx: warning
core/dataplane/rdma_drivers/ofed/src/DtOfedXfer.cxx: warning
core/dataplane/rdma_drivers/pio/src/DtPioXfer.cxx: warning
core/dataplane/rdma_drivers/sockets/src/DtSocketsXfer.cxx: warning
core/dataplane/rdma_smb/linux/src/OcpiPosixFileMappingServices.cxx: warning
core/dataplane/transport/impl/src/OcpiCircuit.cxx: warning
core/dataplane/transport/impl/src/OcpiOutputBuffer.cxx: warning
core/dataplane/transport/impl/src/OcpiPort.cxx: warning
core/hdl/include/HdlAccess.h: string option in getbytes
core/hdl/include/HdlContainer.h: more resilient hdl getMyTicks
core/hdl/include/HdlDevice.h: record device failure, and use it.  add "connect" method
core/hdl/include/HdlNetDriver.h: change Device constructor args
core/hdl/include/HdlOCCP.h: remove dead/unused status bit
core/hdl/include/HdlWciControl.h: do 32 bit reads on < 32 bit properties
core/hdl/src/HdlAccess.cxx: warnings, and string accessors
core/hdl/src/HdlContainer.cxx: better shutdown error handling, verbose info msg about loading
core/hdl/src/HdlDevice.cxx: warning, failure tracking
core/hdl/src/HdlEtherDriver.cxx: move address spaces to different args
core/hdl/src/HdlNetDriver.cxx: ditto
core/hdl/src/HdlPciDriver.cxx: warnings
core/hdl/src/HdlSimDriver.cxx: warnings, and sdp connections
core/hdl/src/HdlSimServer.cxx: warnings, clearer state machine and sdp I/O, major re-do
core/hdl/src/HdlWciControl.cxx: warnings, string access
core/hdl/src/ocfrp_check_main.cxx: warnings
core/hdl/src/ocpihdl_main.cxx: warnings, access arg changes
core/library/src/OcpiLibraryAssembly.cxx: warnings
core/ocl/src/OclContainer.cxx: string accessors
core/rcc/include/RccWorker.h: ditto
core/rcc/src/RccWorker.cxx: ditto
core/remote/src/RemoteDriver.cxx: ditto
core/remote/src/RemoteLauncher.cxx: warnings
core/remote/src/RemoteServer.cxx: warnings
core/util/driver/src/OcpiDriverManager.cxx: more shutdown robustness
core/util/ezxml/include/OcpiUtilEzxml.h: add expression parsing helpers
core/util/ezxml/src/OcpiUtilEzxml.cxx: ditto
core/util/misc/include/OcpiExprEvaluator.h: cleanup
core/util/misc/include/OcpiUtilMisc.h: add fls64, floorlog2 and bitsformax utilities
core/util/misc/src/OcpiUtilMisc.cxx: ditto
core/util/misc/src/OcpiExprEvaluator.cxx: destructor
core/util/property/src/OcpiUtilDataTypes.cxx: warnings, and expression usage
core/util/property/src/OcpiUtilImplementation.cxx: warnings
core/util/property/src/OcpiUtilPValue.cxx: remove this!=NULL test
core/util/property/src/OcpiUtilValue.cxx: warnings
core/util/streamfs/src/OcpiUtilStreamFsReader.cxx: warning
core/util/tcp/src/OcpiUtilTcpStream.cxx: warning
hdl/devices/Makefile: add sdp workers
hdl/platforms/isim_pf/Makefile: add sdp library
hdl/platforms/isim_pf/isim_pf.vhd: use SDP
hdl/platforms/isim_pf/isim_pf.xml: use SDP
hdl/platforms/modelsim_pf/Makefile: use SDP
hdl/platforms/modelsim_pf/modelsim_pf.vhd: use SDP
hdl/platforms/modelsim_pf/modelsim_pf.xml: use SDP
hdl/platforms/specs/platform_spec.xml: have SDP width parameter on platforms
hdl/primitives/Makefile: add sdp
hdl/primitives/ocpi/types_pkg.vhd: add dword_size
hdl/primitives/ocpi/util_body.vhd: another slv to grab LSBs
hdl/primitives/ocpi/util_pkg.vhd: ditto
hdl/primitives/ocpi/wsi_pkg.vhd: add generic s2m and m2s
hdl/primitives/platform/ocscp_rv.vhd: define OCCP_RESET_CLOCKS and use it, other improvements
hdl/primitives/platform/platform_pkg.vhd: make "present" array a bool array
hdl/primitives/util/util_body.vhd: fix some CWD bugs, and better messages, and close_file fcn
hdl/primitives/util/util_pkg.vhd: ditto, and add BRAM2 and more plugarg utilities
jkmac_env_darwin_x86_64.sh: allow c++11 on mac
jkvm6.sh: change license file name
tools/cdk/include/hdl/hdl-assembly.mk: add sdp
tools/cdk/include/hdl/hdl-container.mk: add adp
tools/ocpigen/include/assembly.h: count WTIs and WMEMIs
tools/ocpigen/include/hdl-config.h: allow sdpwidth
tools/ocpigen/include/hdl-container.h: SDP support
tools/ocpigen/include/ocp.h: dwfound, bwfound...
tools/ocpigen/include/port.h: SDPPort support
tools/ocpigen/include/wip.h: ditto
tools/ocpigen/src/assembly.cxx:  count WTIs and WMEMIs
tools/ocpigen/src/data.cxx: data width support
tools/ocpigen/src/hdl-assembly.cxx: sdp_width
tools/ocpigen/src/hdl-config.cxx: sdp_width
tools/ocpigen/src/hdl-container.cxx: sdp support
tools/ocpigen/src/hdl-platform.cxx: warning
tools/ocpigen/src/hdl-worker.cxx: misc
tools/ocpigen/src/hdl.cxx: data width
tools/ocpigen/src/ocp.cxx: remove myfls since it was moved to util, data width
tools/ocpigen/src/parse.cxx: warnings, better data type support for parameters
tools/ocpigen/src/port.cxx: SDP support
tools/ocpigen/src/wci.cxx: use new util functions
tools/ocpigen/src/wmemi.cxx: ditto
tools/ocpigen/src/wmi.cxx: ditto
tools/ocpigen/src/wsi.cxx: ditto
tools/ocpigen/src/wti.cxx: ditto
core/dataplane/rdma_drivers/udp/src/DtUDPSocketXfer.cxx: warning
new SDP files:
core/hdl/include/HdlSdp.h
core/hdl/src/HdlSdp.cxx
hdl/devices/sdp2cp.hdl
hdl/devices/sdp_node.hdl
hdl/devices/sdp_send.hdl
hdl/primitives/sdp/
hdl/primitives/util/plusarg.vhd
tools/ocpigen/src/sdp.cxx

+++ core/util/tcp/src/OcpiUtilTcpStream.cxx
@@ -163,7 +163,7 @@ overflow (int_type i)
   }
 
   char c = traits_type::to_char_type (i);
-  unsigned long long count;
+  size_t count;
 
   try {
     count = m_socket.send (&c, 1);
